iteration 8
refactored code, moving view related code for FoodListScreen in to it's own file FoodListView to create better readability and reduce the length of the FoodListScreen file
