+++ src/components/screens/foodscreens/FoodListScreen.js
@@ -1,21 +1,9 @@
 import { useState, useEffect } from 'react';
-import {
-  LogBox,
-  Text,
-  StyleSheet,
-  View,
-  TouchableOpacity,
-  ScrollView,
-  Alert
-} from 'react-native';
-import Screen from '../../layout/Screen.js';
-import * as Progress from 'react-native-progress';
-import Icons from '../../UI/Icons.js';
-import FoodList from '../../entity/fooditems/FoodList.js';
+import { LogBox, Alert } from 'react-native';
 import { format, addDays, subDays } from 'date-fns';
-import initialFoods from '../../../data/foods.js';
 import { useGoals } from '../../../contexts/GoalsContext';
 import { useMeals } from '../../../contexts/MealsContext';
+import FoodListView from '../../entity/fooditems/FoodListView';
 
 const FoodListScreen = ({ navigation, route }) => {
   // Initialisations -------------------------
@@ -225,254 +213,21 @@ const FoodListScreen = ({ navigation, route }) => {
 
   // View ------------------------------------
   return (
-    <Screen>
-      <View style={styles.dateContainer}>
-        <TouchableOpacity onPress={goToPreviousDay} style={styles.dateButton}>
-          <Text style={styles.dateButtonText}>{'<'}</Text>
-        </TouchableOpacity>
-        <Text style={styles.dateText}>
-          {format(selectedDate, 'EEEE, dd MMM yyyy')}
-        </Text>
-        <TouchableOpacity onPress={goToNextDay} style={styles.dateButton}>
-          <Text style={styles.dateButtonText}>{'>'}</Text>
-        </TouchableOpacity>
-      </View>
-
-      <ScrollView>
-        <View style={styles.headerContainer}>
-          <View style={styles.circleContainer}>
-            <Progress.Circle
-              size={100}
-              progress={Math.min(
-                Object.values(totalCalories).reduce(
-                  (sum, kcal) => sum + kcal,
-                  0
-                ) /
-                  (typeof goals.calories === 'object'
-                    ? goals.calories.max
-                    : goals.calories),
-                1
-              )}
-              showsText={true}
-              color='#E6E6FA'
-              borderWidth={3}
-              textStyle={styles.progressText}
-              formatText={() =>
-                `${Math.round(
-                  Object.values(totalCalories).reduce(
-                    (sum, kcal) => sum + kcal,
-                    0
-                  )
-                )}`
-              }
-            />
-            <Text style={styles.calorieGoalText}>
-              {typeof goals.calories === 'object'
-                ? `${goals.calories.min} - ${goals.calories.max}`
-                : goals.calories}
-            </Text>
-          </View>
-
-          <View style={styles.macrosContainer}>
-            <Text style={styles.macroText}>
-              {Math.round(totalMacros.Protein)}g Protein
-            </Text>
-            <View style={styles.progressBarContainer}>
-              <Progress.Bar
-                progress={Math.min(totalMacros.Protein / goals.protein, 1)}
-                width={150}
-                height={12}
-                color='#E6E6FA'
-                borderRadius={5}
-              />
-              <Text style={styles.progressBarText}>{goals.protein}g</Text>
-            </View>
-
-            <Text style={styles.macroText}>
-              {Math.round(totalMacros.Carbs)}g Carbs
-            </Text>
-            <View style={styles.progressBarContainer}>
-              <Progress.Bar
-                progress={Math.min(totalMacros.Carbs / goals.carbs, 1)}
-                width={150}
-                height={12}
-                color='#E6E6FA'
-                borderRadius={5}
-              />
-              <Text style={styles.progressBarText}>{goals.carbs}g</Text>
-            </View>
-
-            <Text style={styles.macroText}>
-              {Math.round(totalMacros.Fat)}g Fat
-            </Text>
-            <View style={styles.progressBarContainer}>
-              <Progress.Bar
-                progress={Math.min(totalMacros.Fat / goals.fat, 1)}
-                width={150}
-                height={12}
-                color='#E6E6FA'
-                borderRadius={5}
-              />
-              <Text style={styles.progressBarText}>{goals.fat}g</Text>
-            </View>
-          </View>
-        </View>
-        {Object.entries(meals || {}).map(([mealType, foods = []]) => (
-          <View key={mealType} style={styles.mealSection}>
-            <View style={styles.mealHeader}>
-              <Text style={styles.mealTitle}>
-                {mealType}{' '}
-                <Text style={styles.kcalPrMeal}>
-                  {Math.round(totalCalories[mealType] || 0)} kcal
-                </Text>
-              </Text>
-              <View style={styles.mealHeaderButtons}>
-                <TouchableOpacity
-                  onPress={() => handleClearMeal(mealType)}
-                  style={[styles.headerButton, styles.clearButton]}
-                >
-                  <Icons.Delete size={14} />
-                </TouchableOpacity>
-                <TouchableOpacity
-                  onPress={() => gotoAddFood(mealType)}
-                  style={[styles.headerButton, styles.addButton]}
-                >
-                  <Icons.Add />
-                </TouchableOpacity>
-              </View>
-            </View>
-
-            {foods.length === 0 ? (
-              <Text style={styles.dimText}>No food added yet</Text>
-            ) : (
-              <FoodList
-                foods={foods}
-                mealType={mealType}
-                onSelect={(food) => gotoFoodView(food, mealType)}
-                onDelete={handleDeleteFood}
-                onUpdate={handleUpdate}
-              />
-            )}
-          </View>
-        ))}
-      </ScrollView>
-    </Screen>
+    <FoodListView
+      selectedDate={selectedDate}
+      goToPreviousDay={goToPreviousDay}
+      goToNextDay={goToNextDay}
+      totalCalories={totalCalories}
+      goals={goals}
+      totalMacros={totalMacros}
+      meals={meals}
+      handleClearMeal={handleClearMeal}
+      gotoAddFood={gotoAddFood}
+      gotoFoodView={gotoFoodView}
+      handleDeleteFood={handleDeleteFood}
+      handleUpdate={handleUpdate}
+    />
   );
 };
-
-const styles = StyleSheet.create({
-  dateText: {
-    fontSize: 18,
-    fontWeight: 'bold',
-    textAlign: 'center',
-    marginBottom: 10,
-    color: '#665679'
-  },
-  headerContainer: {
-    flexDirection: 'row',
-    justifyContent: 'space-between',
-    alignItems: 'flex-start',
-    paddingHorizontal: 20,
-    paddingVertical: 5
-  },
-  macrosContainer: {
-    paddingTop: -30
-  },
-  progressText: {
-    fontSize: 14,
-    fontWeight: 'semi-bold',
-    color: '#665679'
-  },
-  macroText: {
-    fontSize: 16,
-    marginVertical: 2,
-    color: '#C4C3D0'
-  },
-  mealSection: {
-    marginBottom: 10,
-    paddingHorizontal: 10
-  },
-  mealHeader: {
-    flexDirection: 'row',
-    justifyContent: 'space-between',
-    alignItems: 'center',
-    marginTop: 17,
-    marginBottom: 5
-  },
-  mealTitle: {
-    fontSize: 18,
-    fontWeight: 'bold',
-    color: '#665679'
-  },
-  dimText: {
-    color: '#C4C3D0'
-  },
-  kcalPrMeal: {
-    fontSize: 14,
-    fontWeight: 'light',
-    color: '#C4C3D0'
-  },
-  dateContainer: {
-    flexDirection: 'row',
-    justifyContent: 'space-between',
-    alignItems: 'center',
-    backgroundColor: '#665679',
-    borderRadius: 10,
-    paddingVertical: 3,
-    paddingHorizontal: 10,
-    marginBottom: 15
-  },
-  dateButton: {
-    padding: 10
-  },
-  dateButtonText: {
-    fontSize: 20,
-    color: '#F0EFFF'
-  },
-  dateText: {
-    fontSize: 16,
-    color: '#F0EFFF'
-  },
-  addButton: {
-    backgroundColor: '#F0EFFF'
-  },
-  mealHeaderButtons: {
-    flexDirection: 'row',
-    gap: 8,
-    alignItems: 'center'
-  },
-  headerButton: {
-    width: 25,
-    height: 25,
-    borderRadius: 35 / 2,
-    alignItems: 'center',
-    justifyContent: 'center'
-  },
-  clearButton: {
-    backgroundColor: 'mistyrose'
-  },
-  progressBarContainer: {
-    position: 'relative'
-  },
-  progressBarText: {
-    position: 'absolute',
-    width: '100%',
-    textAlign: 'right',
-    paddingRight: 9,
-    color: '#C4C3D0',
-    fontSize: 9,
-    lineHeight: 13
-  },
-  calorieGoalText: {
-    color: '#665679',
-    fontSize: 14,
-    textAlign: 'center',
-    marginTop: 5
-  },
-  circleContainer: {
-    alignItems: 'center',
-    width: 100
-  }
-});
 
 export default FoodListScreen;
